refactor app to bundles, archives
