+++ isign/archive.py
@@ -1,16 +1,20 @@
+""" Represents an app archive. This is an app at rest, whether it's a naked
+    app bundle in a directory, or a zipped app bundle, or an IPA. We have a
+    common interface to extract these apps to a temp file, then resign them,
+    and create an archive of the same type """
+
 import biplist
-import code_resources
+from bundle import App, is_plist_native
+from exceptions import NotSignable, NotMatched
 from distutils import spawn
-import glob
 import logging
 import os
-from os.path import abspath, basename, exists, isdir, join, splitext
-import re
-from signer import Signer
-import signable
-import shutil
-from subprocess import call
+from os.path import abspath, exists, isdir, join
 import tempfile
+import re
+from subprocess import call
+from signer import Signer
+import shutil
 import zipfile
 
 
@@ -29,154 +33,8 @@ def get_helper(helper_name):
     return helper_paths[helper_name]
 
 
-class NotSignable(Exception):
-    """ superclass for any reason why app shouldn't be
-        signable """
-    pass
-
-
-class NotMatched(NotSignable):
-    """ thrown if we can't find any app class for
-        this file path """
-    pass
-
-
 def make_temp_dir():
     return tempfile.mkdtemp(prefix="isign-")
-
-
-def is_plist_native(plist):
-    return (
-        'CFBundleSupportedPlatforms' in plist and
-        'iPhoneOS' in plist['CFBundleSupportedPlatforms']
-    )
-
-
-class Bundle(object):
-    """ A bundle is a standard directory structure, a signable, installable set of files.
-        Apps are Bundles, but so are some kinds of Frameworks (libraries) """
-    helpers = []
-    executable_class = None
-
-    def __init__(self, path):
-        self.path = path
-        is_native = self.precheck()
-        if not is_native:
-            raise NotMatched("not a native iOS bundle")
-
-        # will be added later
-        self.seal_path = None
-
-        info_path = join(self.path, 'Info.plist')
-        self.info = biplist.readPlist(info_path)
-
-    def precheck(self):
-        """ Checks if a path looks like this kind of app,
-            return stuff we'll need to know about its structure """
-        is_native = False
-        plist_path = join(self.path, "Info.plist")
-        if exists(plist_path):
-            log.debug("got a plist path, {}".format(plist_path))
-            plist = biplist.readPlist(plist_path)
-            is_native = is_plist_native(plist)
-            log.debug("is native: {}".format(is_native))
-        return is_native
-
-    def get_executable_path(self):
-        """ Path to the main executable. For an app, this is app itself. For
-            a Framework, this is the main framework """
-        executable_name = None
-        if 'CFBundleExecutable' in self.info:
-            executable_name = self.info['CFBundleExecutable']
-        else:
-            executable_name, _ = splitext(basename(self.path))
-        executable = join(self.path, executable_name)
-        if not exists(executable):
-            raise Exception(
-                'could not find executable for {0}'.format(self.path))
-        return executable
-
-    def sign(self, signer):
-        """ Sign everything in this bundle, recursively with sub-bundles """
-        log.info("SIGNING: %s" % self.path)
-        frameworks_path = join(self.path, 'Frameworks')
-        if exists(frameworks_path):
-            log.info("SIGNING FRAMEWORKS: %s" % frameworks_path)
-            # sign all the sub-bundles
-            for framework_name in os.listdir(frameworks_path):
-                framework_path = join(frameworks_path, framework_name)
-                log.info("checking for framework: %s" % framework_path)
-                try:
-                    framework = Framework(framework_path)
-                except NotMatched:
-                    log.info("not a framework: %s" % framework_path)
-                    continue
-                log.info("resigning: %s" % framework_path)
-                framework.resign(signer)
-            # sign all the dylibs
-            dylib_paths = glob.glob(join(frameworks_path, '*.dylib'))
-            for dylib_path in dylib_paths:
-                dylib = signable.Dylib(dylib_path)
-                dylib.sign(self, signer)
-        # then create the seal
-        # TODO maybe the app should know what its seal path should be...
-        self.seal_path = code_resources.make_seal(self.get_executable_path(),
-                                                  self.path)
-        # then sign the app
-        executable = self.executable_class(self.get_executable_path())
-        executable.sign(self, signer)
-
-    def resign(self, signer):
-        """ signs app, modifies appdir in place """
-        self.sign(signer)
-        log.debug("Resigned app dir at <%s>", self.path)
-
-
-class Framework(Bundle):
-    """ A bundle that comprises reusable code. Similar to an app in that it has
-        its own resources and metadata. Not like an app because the main executable
-        doesn't have Entitlements, or an Application hash, and it doesn't have its
-        own provisioning profile. """
-
-    # the executable in this bundle will be a Framework
-    executable_class = signable.Framework
-
-    def __init__(self, path):
-        super(Framework, self).__init__(path)
-
-
-class App(Bundle):
-    """ The kind of bundle that is visible as an app to the user.
-        Contains the provisioning profile, entitlements, etc.  """
-
-    # the executable in this bundle will be an Executable (i.e. an App)
-    executable_class = signable.Executable
-
-    def __init__(self, path):
-        super(App, self).__init__(path)
-        self.entitlements_path = join(self.path,
-                                      'Entitlements.plist')
-        self.provision_path = join(self.path,
-                                   'embedded.mobileprovision')
-
-    def provision(self, provision_path):
-        shutil.copyfile(provision_path, self.provision_path)
-
-    def create_entitlements(self, team_id):
-        entitlements = {
-            "keychain-access-groups": [team_id + '.*'],
-            "com.apple.developer.team-identifier": team_id,
-            "application-identifier": team_id + '.*',
-            "get-task-allow": True
-        }
-        biplist.writePlist(entitlements, self.entitlements_path, binary=False)
-        log.debug("wrote Entitlements to {0}".format(self.entitlements_path))
-
-    def resign(self, signer, provisioning_profile):
-        """ signs app, modifies appdir in place """
-        self.provision(provisioning_profile)
-        self.create_entitlements(signer.team_id)
-        super(App, self).resign(signer)
 
 
 class AppArchive(object):
@@ -207,6 +65,17 @@ class AppZip(object):
     extensions = ['.zip']
     helpers = ['zip', 'unzip']
 
+    def __init__(self, path):
+        self.path = path
+        if not self.is_helpers_present():
+            raise NotSignable("helpers not present")
+        relative_app_dir, is_native = self.precheck()
+        self.relative_app_dir = relative_app_dir
+        if relative_app_dir is None:
+            raise NotMatched("no app directory found")
+        if not is_native:
+            raise NotMatched("not a native iOS app")
+
     def is_helpers_present(self):
         """ returns False if any of our helper apps wasn't found in class init """
         is_present = True
@@ -226,7 +95,8 @@ class AppZip(object):
 
     def precheck(self):
         """ Checks if an archive looks like this kind of app. Have to examine
-            within the zipfile, b/c we don't want to make temp dirs just yet """
+            within the zipfile, b/c we don't want to make temp dirs just yet. This
+            recapitulates a very similar precheck in the Bundle class """
         relative_app_dir = None
         is_native = False
         if (self.is_archive_extension_match() and
@@ -248,17 +118,6 @@ class AppZip(object):
                 log.debug("is_native: {}".format(is_native))
         return (relative_app_dir, is_native)
 
-    def __init__(self, path):
-        self.path = path
-        if not self.is_helpers_present():
-            raise NotSignable("helpers not present")
-        relative_app_dir, is_native = self.precheck()
-        self.relative_app_dir = relative_app_dir
-        if relative_app_dir is None:
-            raise NotMatched("no app directory found")
-        if not is_native:
-            raise NotMatched("not a native iOS app")
-
     def unarchive_to_temp(self):
         containing_dir = make_temp_dir()
         call([get_helper('unzip'), "-qu", self.path, "-d", containing_dir])
@@ -267,6 +126,8 @@ class AppZip(object):
 
     @classmethod
     def archive(cls, containing_dir, output_path):
+        """ archive this up into a zipfile. Note this is a classmethod, because
+            the caller will use us on a temp directory somewhere """
         # the temp file is necessary because zip always adds ".zip" if it
         # does not have an extension. But we want to respect the desired
         # output_path's extension, which could be ".ipa" or who knows.
@@ -301,7 +162,8 @@ class Ipa(AppZip):
 
 
 def archive_factory(path):
-    """ factory to unpack various app types. """
+    """ Guess what kind of archive we are dealing with, return an
+        archive object. """
     if isdir(path):
         try:
             return AppArchive(path)
@@ -329,6 +191,9 @@ def resign(input_path,
            apple_cert,
            provisioning_profile,
            output_path):
+    """ Unified interface to extract any kind of archive from
+        a temporary file, resign it with these credentials,
+        and create a similar archive for that resigned app """
 
     if not exists(input_path):
         raise IOError("{0} not found".format(input_path))
@@ -340,8 +205,8 @@ def resign(input_path,
     temp_dir = None
     try:
         archive = archive_factory(input_path)
-        (temp_dir, app) = archive.unarchive_to_temp()
-        app.resign(signer, provisioning_profile)
+        (temp_dir, bundle) = archive.unarchive_to_temp()
+        bundle.resign(signer, provisioning_profile)
         archive.__class__.archive(temp_dir, output_path)
     except NotSignable as e:
         msg = "Not signable: <{0}>: {1}\n".format(input_path, e)
