+++ tests/generate_codesig_construct_txt.py
@@ -2,7 +2,7 @@
 # test files such as Test.app.codesig.construct.txt
 
 from isign_base_test import IsignBaseTest
-import isign.app
+import isign.archive
 from isign.signable import Executable
 
 import logging
@@ -21,7 +21,7 @@ def log_to_stderr(level=logging.INFO):
 log_to_stderr(logging.DEBUG)
 log = logging.getLogger(__name__)
 log.debug("generating a signature parse for an app")
-app = isign.app.App(IsignBaseTest.TEST_APP)
+app = isign.archive.AppArchive(IsignBaseTest.TEST_APP)
 executable = Executable(app.get_executable_path())
 arch = executable.arches[0]
 codesig_str = str(arch['cmds']['LC_CODE_SIGNATURE'])
