+++ tests/test_parsing.py
@@ -1,5 +1,5 @@
 from isign_base_test import IsignBaseTest
-import isign.app
+import isign.bundle
 from isign.signable import Executable
 import logging
 
@@ -16,8 +16,8 @@ class TestParsing(IsignBaseTest):
     def test_app(self):
         with open(self.TEST_APP_CODESIG_STR, 'r') as f:
             expected_codesig_str = f.read().strip()
-        app = isign.app.App(self.TEST_APP)
-        executable = Executable(app.get_executable_path())
+        bundle = isign.bundle.App(self.TEST_APP)
+        executable = Executable(bundle.get_executable_path())
         arch = executable.arches[0]
         codesig_str = str(arch['cmds']['LC_CODE_SIGNATURE'])
         self.assertEquals(expected_codesig_str, codesig_str)
